requirements.txt
python-docx>=1.1.0
pdfplumber>=0.11.0
pytest>=8.1.1
